WIP feat(expense): partial Decimal migration — cascade detected, scope L+ (EXP-003)
⚠️ **Cette commit ne compile pas — cascade plus large que prévu**.

Story EXP-003 du sprint W18 (#430) a sous-estimé l'ampleur du blast radius.
La migration f64 → Decimal sur expense.rs + invoice_line_item.rs + expense_dto.rs
casse en cascade ~7+ fichiers downstream qui consomment ces types comme f64.

## Migré (compile OK en isolation)

- backend/src/domain/entities/expense.rs : 8 f64 → Decimal, struct + 2 ctors + recalculate_vat + tous les tests existants migrés via dec!()
- backend/src/domain/entities/invoice_line_item.rs : 12 f64 → Decimal, full rewrite avec 18 tests 4-cat (3 happy + 7 edge + 4 security + 4 negative) incl. test critique edge_decimal_exactness_preserved_on_cumul (valide 0.1 + 0.2 == 0.3)
- backend/src/application/dto/expense_dto.rs : 22 f64 → Decimal sur DTOs (CreateExpense, ExpenseResponse, CreateInvoiceDraft, UpdateInvoiceDraft, CreateInvoiceLineItem, InvoiceResponse, InvoiceLineItemResponse, ChargeDistributionResponse)
- backend/Cargo.toml : promote rust_decimal_macros from [dev-dependencies] → [dependencies] (dec!() macro maintenant utilisé en code production)

## Cascade non migrée (36 erreurs cargo check)

Modules downstream qui consomment expense.amount / invoice_line_item fields comme f64 :

- src/application/services/expense_accounting_service.rs (~6 erreurs : journal_entry mul/sum)
- src/application/use_cases/charge_distribution_use_cases.rs (1)
- src/application/use_cases/dashboard_use_cases.rs (~8 erreurs : sums Iterator<f64>)
- src/application/use_cases/expense_use_cases.rs
- src/application/use_cases/budget_use_cases.rs
- src/application/use_cases/payment_reminder_use_cases.rs
- src/application/dto/payment_reminder_dto.rs (consume Expense)
- src/infrastructure/database/repositories/expense_repository_impl.rs (sqlx mapping)
- src/infrastructure/web/handlers/expense_handlers.rs (~2 erreurs)
- src/infrastructure/web/handlers/building_handlers.rs (1, sum)
- src/infrastructure/web/handlers/mcp_sse_handlers.rs (~2+ erreurs : sums)

Pattern : ~36 sites de mismatched types ou Iterator<Decimal>::sum::<f64>().

## Pourquoi ce commit malgré la non-compilation ?

Discipline Maury : **rendre le problème visible publiquement plutôt que le cacher**.
Cette WIP commit + PR draft documentent honnêtement que la story EXP-003 telle
que définie dans #430 et #433 est sous-estimée (M annoncé, L+ effectif).

Plutôt que :
- (a) push à travers et migrer 50+ sites en un seul gros PR (risque review impossible)
- (b) revenir en arrière (perd l'investissement entité+DTO solide)
- (c) ajouter des conversions `.to_f64()` partout (anti-pattern explicit
  ADR-0007 — leak la précision à chaque boundary, défait l'objectif)

→ surface le scope réel et propose un découpage propre en stories follow-up.

## Découpage proposé pour finir la migration

**EXP-003a (cette story, à compléter)** : entité expense + invoice_line_item + DTO ✅
                                          (tests 4-cat en isolation passeront
                                          quand cascade fixée)

**EXP-003b (NEW)** : services + use_cases consommant Expense
  - expense_accounting_service.rs
  - charge_distribution_use_cases.rs
  - expense_use_cases.rs
  - budget_use_cases.rs (juste passe-through)
  - payment_reminder_use_cases.rs + payment_reminder_dto.rs

**EXP-003c (NEW)** : repositories + handlers
  - expense_repository_impl.rs (sqlx mapping vers NUMERIC — cf. ADR-0008)
  - expense_handlers.rs
  - + migration SQL alter_expense_amounts_to_numeric

**EXP-003d (NEW)** : modules périphériques avec sum()
  - dashboard_use_cases.rs (calculs agrégats Decimal au lieu de f64)
  - building_handlers.rs (sum expenses)
  - mcp_sse_handlers.rs (sums dans MCP responses)

Total : ~3-4 stories M chacune, ~1 semaine de travail. Sprint W19 minimum.

Refs: EXP-002 (audit) #434 ADR-decimal-policy #435 umbrella #433 sprint #430
      ADR-0007 (Decimal Rust) ADR-0008 (NUMERIC PG) ADR-0009 (IoT keep f64)

+++ backend/src/application/dto/expense_dto.rs
@@ -1,4 +1,12 @@
+//! Expense DTOs — monetary fields use `rust_decimal::Decimal` (cf. ADR-0007).
+//!
+//! Note : `validator` crate `#[validate(range(...))]` ne support pas Decimal
+//! avec literals de type f64. Les invariants montants (> 0, taux VAT 0-100)
+//! sont enforced dans `Expense::new` / `Expense::new_with_vat` côté domaine
+//! (cf. `domain/entities/expense.rs`).
+
 use crate::domain::entities::{ApprovalStatus, ExpenseCategory, PaymentStatus};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use validator::Validate;
 
@@ -14,8 +22,8 @@ pub struct CreateExpenseDto {
     #[validate(length(min = 1))]
     pub description: String,
 
-    #[validate(range(min = 0.01))]
-    pub amount: f64,
+    /// Montant TTC (validé > 0 dans Expense::new). Decimal exact (cf. ADR-0007).
+    pub amount: Decimal,
 
     pub expense_date: String,
     pub supplier: Option<String>,
@@ -33,7 +41,7 @@ pub struct ExpenseResponseDto {
     pub building_id: String,
     pub category: ExpenseCategory,
     pub description: String,
-    pub amount: f64,
+    pub amount: Decimal,
     pub expense_date: String,
     pub payment_status: PaymentStatus,
     pub approval_status: ApprovalStatus,
@@ -47,7 +55,8 @@ pub struct ExpenseResponseDto {
 
 // ========== New Invoice DTOs (with VAT & Workflow) ==========
 
-/// Créer une facture brouillon avec gestion TVA
+/// Créer une facture brouillon avec gestion TVA.
+/// Validation des montants > 0 et taux 0-100 effectuée dans `Expense::new_with_vat`.
 #[derive(Debug, Deserialize, Validate, Clone)]
 pub struct CreateInvoiceDraftDto {
     #[serde(default)]
@@ -58,11 +67,11 @@ pub struct CreateInvoiceDraftDto {
     #[validate(length(min = 1))]
     pub description: String,
 
-    #[validate(range(min = 0.01))]
-    pub amount_excl_vat: f64, // Montant HT
+    /// Montant HT (validé > 0 dans `Expense::new_with_vat`).
+    pub amount_excl_vat: Decimal,
 
-    #[validate(range(min = 0.0, max = 100.0))]
-    pub vat_rate: f64, // Taux TVA (21.0 pour 21%)
+    /// Taux TVA en % (validé 0..=100 dans `Expense::new_with_vat`).
+    pub vat_rate: Decimal,
 
     pub invoice_date: String,     // ISO 8601
     pub due_date: Option<String>, // ISO 8601
@@ -70,7 +79,7 @@ pub struct CreateInvoiceDraftDto {
     pub invoice_number: Option<String>,
 }
 
-/// Modifier une facture brouillon ou rejetée
+/// Modifier une facture brouillon ou rejetée.
 #[derive(Debug, Deserialize, Validate, Clone)]
 pub struct UpdateInvoiceDraftDto {
     #[validate(length(min = 1))]
@@ -78,11 +87,8 @@ pub struct UpdateInvoiceDraftDto {
 
     pub category: Option<ExpenseCategory>,
 
-    #[validate(range(min = 0.01))]
-    pub amount_excl_vat: Option<f64>,
-
-    #[validate(range(min = 0.0, max = 100.0))]
-    pub vat_rate: Option<f64>,
+    pub amount_excl_vat: Option<Decimal>,
+    pub vat_rate: Option<Decimal>,
 
     pub invoice_date: Option<String>,
     pub due_date: Option<String>,
@@ -90,19 +96,19 @@ pub struct UpdateInvoiceDraftDto {
     pub invoice_number: Option<String>,
 }
 
-/// Soumettre une facture pour validation (Draft → PendingApproval)
+/// Soumettre une facture pour validation (Draft → PendingApproval).
 #[derive(Debug, Deserialize, Clone)]
 pub struct SubmitForApprovalDto {
     // Empty body, action via PUT /invoices/:id/submit
 }
 
-/// Approuver une facture (PendingApproval → Approved)
+/// Approuver une facture (PendingApproval → Approved).
 #[derive(Debug, Deserialize, Clone)]
 pub struct ApproveInvoiceDto {
     pub approved_by_user_id: String, // User ID du syndic/admin
 }
 
-/// Rejeter une facture avec raison (PendingApproval → Rejected)
+/// Rejeter une facture avec raison (PendingApproval → Rejected).
 #[derive(Debug, Deserialize, Validate, Clone)]
 pub struct RejectInvoiceDto {
     pub rejected_by_user_id: String,
@@ -111,7 +117,8 @@ pub struct RejectInvoiceDto {
     pub rejection_reason: String,
 }
 
-/// Créer une ligne de facture
+/// Créer une ligne de facture.
+/// Validations (quantity > 0, unit_price ≥ 0, vat_rate 0..=100) dans `InvoiceLineItem::new`.
 #[derive(Debug, Deserialize, Validate, Clone)]
 pub struct CreateInvoiceLineItemDto {
     pub expense_id: String,
@@ -119,19 +126,14 @@ pub struct CreateInvoiceLineItemDto {
     #[validate(length(min = 1))]
     pub description: String,
 
-    #[validate(range(min = 0.01))]
-    pub quantity: f64,
-
-    #[validate(range(min = 0.0))]
-    pub unit_price: f64,
-
-    #[validate(range(min = 0.0, max = 100.0))]
-    pub vat_rate: f64,
+    pub quantity: Decimal,
+    pub unit_price: Decimal,
+    pub vat_rate: Decimal,
 }
 
 // ========== Response DTOs ==========
 
-/// Response enrichie avec tous les champs invoice/workflow
+/// Response enrichie avec tous les champs invoice/workflow.
 #[derive(Debug, Serialize, Clone)]
 pub struct InvoiceResponseDto {
     pub id: String,
@@ -140,12 +142,12 @@ pub struct InvoiceResponseDto {
     pub category: ExpenseCategory,
     pub description: String,
 
-    // Montants
-    pub amount: f64, // TTC (backward compatibility)
-    pub amount_excl_vat: Option<f64>,
-    pub vat_rate: Option<f64>,
-    pub vat_amount: Option<f64>,
-    pub amount_incl_vat: Option<f64>,
+    // Montants — exact decimal (cf. ADR-0007)
+    pub amount: Decimal, // TTC (backward compatibility)
+    pub amount_excl_vat: Option<Decimal>,
+    pub vat_rate: Option<Decimal>,
+    pub vat_amount: Option<Decimal>,
+    pub amount_incl_vat: Option<Decimal>,
 
     // Dates
     pub expense_date: String,
@@ -172,34 +174,35 @@ pub struct InvoiceResponseDto {
     pub updated_at: String,
 }
 
-/// Response pour une ligne de facture
+/// Response pour une ligne de facture.
 #[derive(Debug, Serialize)]
 pub struct InvoiceLineItemResponseDto {
     pub id: String,
     pub expense_id: String,
     pub description: String,
-    pub quantity: f64,
-    pub unit_price: f64,
-    pub amount_excl_vat: f64,
-    pub vat_rate: f64,
-    pub vat_amount: f64,
-    pub amount_incl_vat: f64,
+    pub quantity: Decimal,
+    pub unit_price: Decimal,
+    pub amount_excl_vat: Decimal,
+    pub vat_rate: Decimal,
+    pub vat_amount: Decimal,
+    pub amount_incl_vat: Decimal,
     pub created_at: String,
 }
 
-/// Response pour une répartition de charge
+/// Response pour une répartition de charge.
 #[derive(Debug, Serialize)]
 pub struct ChargeDistributionResponseDto {
     pub id: String,
     pub expense_id: String,
     pub unit_id: String,
     pub owner_id: String,
-    pub quota_percentage: f64, // Sera converti en pourcentage côté client (0.25 → 25%)
-    pub amount_due: f64,
+    /// Quote-part (e.g., dec!(0.25) pour 25%). Decimal exact (cf. ADR-0007).
+    pub quota_percentage: Decimal,
+    pub amount_due: Decimal,
     pub created_at: String,
 }
 
-/// Liste des factures en attente d'approbation (pour syndics)
+/// Liste des factures en attente d'approbation (pour syndics).
 #[derive(Debug, Serialize)]
 pub struct PendingInvoicesListDto {
     pub invoices: Vec<InvoiceResponseDto>,
